Добавил вычисление общей стоимости и вынес функции в модуль utils
