+++ public/index.html
@@ -19,9 +19,7 @@
           <section class="trip-main__trip-info  trip-info">
             <!-- Маршрут -->
 
-            <p class="trip-info__cost">
-              Total: &euro;&nbsp;<span class="trip-info__cost-value">0</span>
-            </p>
+
           </section>
 
           <div class="trip-main__trip-controls  trip-controls">
