+++ src/components/trip.js
@@ -1,5 +1,8 @@
-export const getTripHTML = (title = ``, dates = ``) => `
+export const getTripHTML = (title = ``, dates = ``, cost = 0) => `
   <div class="trip-info__main">
     <h1 class="trip-info__title">${title}</h1>
     <p class="trip-info__dates">${dates}</p>
-  </div>`;
+  </div>
+  <p class="trip-info__cost">
+    Total: &euro;&nbsp;<span class="trip-info__cost-value">${cost}</span>
+  </p>`;
