+++ src/main.js
@@ -8,43 +8,21 @@ import {getEventEditHTML} from './components/event-edit';
 import {getEventItemHTML} from './components/event-item';
 import {getEvent} from './data/event';
 
+import {where, renderElem, uniqueArray, getTotalCost} from './utils';
+
 (function () {
-  // consts for renderElem function (values for param "place")
-  const where = {
-    beforeBegin: `beforeBegin`,
-    afterBegin: `afterBegin`,
-    beforeEnd: `beforeEnd`,
-    afterEnd: `afterEnd`
-  };
-
-  // render element function
-  const renderElem = function (elem, htmlCode, place = where.beforeEnd) {
-    elem.insertAdjacentHTML(place, htmlCode);
-  };
-
-  // get unique elements of array
-  function uniqueArray(arr) {
-    let result = [];
-
-    for (let str of arr) {
-      if (!result.includes(str)) {
-        result.push(str);
-      }
-    }
-
-    return result;
-  }
-
-  const elemTripMain = document.querySelector(`.page-header .trip-main`);
+  const elemTripMain = document.querySelector(`.trip-main`);
   const elemTripInfo = elemTripMain.querySelector(`.trip-info`);
   const elemTripControls = elemTripMain.querySelector(`.trip-controls`);
   const elemTripControlsH = elemTripControls.querySelector(`h2.visually-hidden`);
-
   const elemPageMain = document.querySelector(`.page-main`);
   const elemTripEvents = elemPageMain.querySelector(`.trip-events`);
 
+  const EVENT_COUNT = 10;
+  const arrTripEvents = new Array(EVENT_COUNT).fill().map(getEvent);
+
   // trip
-  renderElem(elemTripInfo, getTripHTML(`Amsterdam &mdash; ... &mdash; Amsterdam`, `Mar 18&nbsp;&mdash;&nbsp;22`), where.afterBegin);
+  renderElem(elemTripInfo, getTripHTML(`Amsterdam &mdash; ... &mdash; Amsterdam`, `Mar 18&nbsp;&mdash;&nbsp;22`, getTotalCost(arrTripEvents)));
 
   // menu
   renderElem(elemTripControlsH, getMenuHTML(), where.afterEnd);
@@ -59,15 +37,8 @@ import {getEvent} from './data/event';
   const tripDays = document.createElement(`ul`);
   tripDays.className = `trip-days`;
 
-
-  const EVENT_COUNT = 10;
-  const arrTripEvents = new Array(EVENT_COUNT).fill().map(getEvent);
-
-  window.arrTripEvents = arrTripEvents;
-
   // array of trip days
   const arrTripDays = uniqueArray(arrTripEvents.map((event) => (new Date(event.dateBegin).setHours(0, 0, 0, 0)))).sort();
-
 
   // days
   arrTripDays.forEach((day, dayItndex) => {
